pyproject.toml
[tool.poetry]
name = "python-web-api"
version = "1.0.0"
description = "A Web API Reference"
authors = ["Randal Vance Cunanan <randal@randalvance.net>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.88.0"}
sqlmodel = "^0.0.8"
psycopg2-binary = "^2.9.5"
pyjwt = {extras = ["crypto"], version = "^2.6.0"}


[tool.poetry.group.dev.dependencies]
alembic = "^1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
